removing unnecessary commits

+++ server/src/test/java/org/elasticsearch/cluster/ClusterStateTests.java
@@ -1153,7 +1153,7 @@ public class ClusterStateTests extends ESTestCase {
     public static int expectedChunkCount(ToXContent.Params params, ClusterState clusterState) {
         final var metrics = ClusterState.Metric.parseString(params.param("metric", "_all"), true);
 
-        long chunkCount = 0;
+        int chunkCount = 0;
 
         // header chunk
         chunkCount += 1;
@@ -1217,6 +1217,6 @@ public class ClusterStateTests extends ESTestCase {
             }
         }
 
-        return (int) chunkCount;
+        return chunkCount;
     }
 }
